requirements.txt
# django==2.2
pytesseract==0.3.10
Pillow==9.5.0